+++ TuDienAnhViet/Program.cs
@@ -16,7 +16,7 @@ namespace TuDienAnhViet
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new MainForm());
+            Application.Run(new FormSearch());
         }
     }
 }
